feat(edit-content-bridge) implement a factory and bridge pattern
