+++ core-web/libs/edit-content-bridge/src/index.ts
@@ -1,1 +1,3 @@
-export { DotFormBridge } from './lib/edit-content-bridge';
+export { createFormBridge } from './lib/factories/form-bridge.factory';
+
+export { FormBridge } from './lib/interfaces/form-bridge.interface';
